Integrated login and logout into the web app, hides navbar, implemented auth guards by route
