+++ src/client/src/app/app-routing.module.ts
@@ -1,82 +1,85 @@
-import { NgModule }                 from '@angular/core';
-import { Routes, RouterModule }     from "@angular/router";
-import { LoginComponent }           from "./page/login/login.component";
+import { NgModule }                  from '@angular/core';
+import { Routes, RouterModule }      from "@angular/router";
+import { LoginOldComponent }         from "./page/login-old/login-old.component";
 import { UserRegistrationComponent } from "./page/user-registration/user-registration.component";
-import { NotFoundComponent }        from "./page/not-found/not-found.component";
-import { HomeComponent }            from "./page/home/home.component";
-import { LogoutComponent }          from "./page/logout/logout.component";
-import {SettingsComponent }         from "./page/settings/settings.component";
-import {SalesComponent }            from "./page/sales/sales.component";
-import { UsersComponent }           from "./page/users/users.component";
-import { InventoryComponent }       from "./page/inventory/inventory.component";
+import { NotFoundComponent }         from "./page/not-found/not-found.component";
+import { HomeComponent }             from "./page/home/home.component";
+import { LogoutComponent }           from "./page/logout/logout.component";
+import { SettingsComponent }         from "./page/settings/settings.component";
+import { SalesComponent }            from "./page/sales/sales.component";
+import { UsersComponent }            from "./page/users/users.component";
+import { InventoryComponent }        from "./page/inventory/inventory.component";
+import { LoginComponent }            from "./page/login/login.component";
+import { AuthGuard }                 from "./guard/auth.guard";
 
 const routes: Routes = [
 
   {
     path         : 'home',
     component    : HomeComponent,
-    data         : { title: "Point of Sale"}
+    data         : { title: "Point of Sale"},
+    canActivate  : [AuthGuard]
   },
-
   {
     path         : 'inventory',
     component    : InventoryComponent,
-    data         : { title: "Inventory"}
+    data         : { title: "Inventory"},
+    canActivate  : [AuthGuard]
   },
-
   {
     path         : 'users',
     component    : UsersComponent,
-    data         : { title: "Users" }
-
+    data         : { title: "Users" },
+    canActivate  : [AuthGuard]
   },
-
   {
     path         : "",
     redirectTo   : "home",
     pathMatch    : "full"
   },
-
   {
-    path          : 'login',
-    component     : LoginComponent,
-    data          : { title: 'Login' }
+    path          : 'login-old',
+    component     : LoginOldComponent,
+    data          : { title: 'Old Login' }
   },
-
   {
     path          : 'logout',
     component     : LogoutComponent,
     data          : { title: 'Logout' }
   },
-
   {
     path         : 'settings',
     component    : SettingsComponent,
-    data         : { title: "Settings"}
+    data         : { title: "Settings"},
+    canActivate  : [AuthGuard]
   },
-
   {
     path         : 'sales',
     component    : SalesComponent,
-    data         : { title: "Sales"}
+    data         : { title: "Sales"},
+    canActivate  : [AuthGuard]
+  },
+  {
+    path      : 'login',
+    component : LoginComponent,
+    data      : { title: "Login"}
+  },
+  {
+    path        : 'registration',
+    component   : UserRegistrationComponent,
+    data        : { title: 'User Registration' },
+    canActivate : [AuthGuard]
+  },
+  {
+    path      : 'not-found',
+    component : NotFoundComponent,
+    data      : { title: 'Page Not Found' }
   },
 
   {
-    path          : 'registration',
-    component     : UserRegistrationComponent,
-    data          : { title: 'User Registration' }
-  },
-
-  {
-    path          : 'not-found',
-    component     : NotFoundComponent,
-    data          : { title: 'Page Not Found' }
-  },
-
-  { path: '**',
+    path: '**',
     redirectTo: 'not-found'
   }
-
 ];
 
 @NgModule( {
@@ -86,8 +89,10 @@ const routes: Routes = [
 })
 
 export class AppRoutingModule { }
+
 export const routingComponents = [
-  LoginComponent,
+
+  LoginOldComponent,
   UserRegistrationComponent,
   SalesComponent,
   HomeComponent,
@@ -95,5 +100,6 @@ export const routingComponents = [
   NotFoundComponent,
   LogoutComponent,
   UsersComponent,
-  InventoryComponent
+  InventoryComponent,
+  LoginComponent
 ];
