+++ src/client/src/app/page/home/home.component.ts
@@ -5,11 +5,13 @@ import { Component, OnInit } from '@angular/core';
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.css']
 })
+
 export class HomeComponent implements OnInit {
 
   constructor() { }
 
   ngOnInit() {
+
   }
 
 }
