+++ src/client/src/app/page/login-old/login-old.component.html
@@ -1,0 +1,73 @@
+<div class="alert alert-danger" id="notLoggedIn">Not logged in!</div>
+
+<div class="row">
+  <div class="col-md-6">
+    <div class="panel panel-default" id="login">
+      <div class="panel-heading">
+        <h3 class="panel-title">Login</h3>
+      </div>
+      <div class="panel-body">
+        <form id="loginForm">
+          <div class="form-group">
+            <input type="text" class="form-control" id="exampleInputEmail1" placeholder="username"
+                   required name="username">
+          </div>
+          <div class="form-group">
+            <input type="password" class="form-control" id="exampleInputPassword1"
+                   placeholder="password" required name="password">
+          </div>
+          <div class="well">
+            Try one of the following logins
+            <ul>
+              <li>admin & admin</li>
+              <li>user & password</li>
+              <li>disabled & password</li>
+            </ul>
+          </div>
+          <button type="submit" class="btn btn-primary">login</button>
+        </form>
+      </div>
+    </div>
+
+    <div id="userInfo">
+      <div class="panel panel-default">
+        <div class="panel-heading">
+          <h3 class="panel-title">Authenticated user</h3>
+        </div>
+        <div class="panel-body">
+          <div id="userInfoBody"></div>
+          <button type="button" class="btn btn-danger" id="logoutButton">logout</button>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="col-md-6">
+    <div class="btn-group" role="group" aria-label="..." style="margin-bottom: 16px;">
+      <button type="button" class="btn btn-outline-primary" id="exampleServiceBtn">call example service</button>
+      <button type="button" class="btn btn-outline-primary" id="adminServiceBtn">call admin protected service</button>
+    </div>
+
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3 class="panel-title">Response:</h3>
+      </div>
+      <div class="panel-body">
+        <pre id="response"></pre>
+      </div>
+    </div>
+  </div>
+</div>
+
+
+<div class="row">
+  <div id="loggedIn" class="col-md-6">
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3 class="panel-title">Token information</h3>
+      </div>
+      <div class="panel-body" id="loggedInBody"></div>
+    </div>
+  </div>
+</div>
+
