+++ src/client/src/app/page/login/login.component.html
@@ -1,73 +1,37 @@
-<div class="alert alert-danger" id="notLoggedIn">Not logged in!</div>
+<app-card-page>
 
-<div class="row">
-  <div class="col-md-6">
-    <div class="panel panel-default" id="login">
-      <div class="panel-heading">
-        <h3 class="panel-title">Login</h3>
-      </div>
-      <div class="panel-body">
-        <form id="loginForm">
+  <div class="container">
+    <div class="row">
+      <div class="col-md-6 offset-md-3">
+
+        <h2 class="center">Login</h2>
+        <form [formGroup]="loginForm" (ngSubmit)="onSubmit()">
           <div class="form-group">
-            <input type="text" class="form-control" id="exampleInputEmail1" placeholder="username"
-                   required name="username">
+            <label for="username">Username</label>
+            <input type="text" formControlName="username" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.username.errors }" />
+            <div *ngIf="submitted && f.username.errors" class="invalid-feedback">
+              <div *ngIf="f.username.errors.required">Username is required</div>
+            </div>
           </div>
+
+
           <div class="form-group">
-            <input type="password" class="form-control" id="exampleInputPassword1"
-                   placeholder="password" required name="password">
+            <label for="password">Password</label>
+            <input type="password" formControlName="password" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.password.errors }" />
+            <div *ngIf="submitted && f.password.errors" class="invalid-feedback">
+              <div *ngIf="f.password.errors.required">Password is required</div>
+            </div>
           </div>
-          <div class="well">
-            Try one of the following logins
-            <ul>
-              <li>admin & admin</li>
-              <li>user & password</li>
-              <li>disabled & password</li>
-            </ul>
+
+
+          <div class="form-group center">
+            <button [disabled]="loading" class="btn btn-primary">Login</button>
+            <img *ngIf="loading" src="data:image/gif;base64,R0lGODlhEAAQAPIAAP///wAAAMLCwkJCQgAAAGJiYoKCgpKSkiH/C05FVFNDQVBFMi4wAwEAAAAh/hpDcmVhdGVkIHdpdGggYWpheGxvYWQuaW5mbwAh+QQJCgAAACwAAAAAEAAQAAADMwi63P4wyklrE2MIOggZnAdOmGYJRbExwroUmcG2LmDEwnHQLVsYOd2mBzkYDAdKa+dIAAAh+QQJCgAAACwAAAAAEAAQAAADNAi63P5OjCEgG4QMu7DmikRxQlFUYDEZIGBMRVsaqHwctXXf7WEYB4Ag1xjihkMZsiUkKhIAIfkECQoAAAAsAAAAABAAEAAAAzYIujIjK8pByJDMlFYvBoVjHA70GU7xSUJhmKtwHPAKzLO9HMaoKwJZ7Rf8AYPDDzKpZBqfvwQAIfkECQoAAAAsAAAAABAAEAAAAzMIumIlK8oyhpHsnFZfhYumCYUhDAQxRIdhHBGqRoKw0R8DYlJd8z0fMDgsGo/IpHI5TAAAIfkECQoAAAAsAAAAABAAEAAAAzIIunInK0rnZBTwGPNMgQwmdsNgXGJUlIWEuR5oWUIpz8pAEAMe6TwfwyYsGo/IpFKSAAAh+QQJCgAAACwAAAAAEAAQAAADMwi6IMKQORfjdOe82p4wGccc4CEuQradylesojEMBgsUc2G7sDX3lQGBMLAJibufbSlKAAAh+QQJCgAAACwAAAAAEAAQAAADMgi63P7wCRHZnFVdmgHu2nFwlWCI3WGc3TSWhUFGxTAUkGCbtgENBMJAEJsxgMLWzpEAACH5BAkKAAAALAAAAAAQABAAAAMyCLrc/jDKSatlQtScKdceCAjDII7HcQ4EMTCpyrCuUBjCYRgHVtqlAiB1YhiCnlsRkAAAOwAAAAAAAAAAAA==" />
           </div>
-          <button type="submit" class="btn btn-primary">login</button>
         </form>
-      </div>
-    </div>
 
-    <div id="userInfo">
-      <div class="panel panel-default">
-        <div class="panel-heading">
-          <h3 class="panel-title">Authenticated user</h3>
-        </div>
-        <div class="panel-body">
-          <div id="userInfoBody"></div>
-          <button type="button" class="btn btn-danger" id="logoutButton">logout</button>
-        </div>
+
       </div>
     </div>
   </div>
-
-  <div class="col-md-6">
-    <div class="btn-group" role="group" aria-label="..." style="margin-bottom: 16px;">
-      <button type="button" class="btn btn-outline-primary" id="exampleServiceBtn">call example service</button>
-      <button type="button" class="btn btn-outline-primary" id="adminServiceBtn">call admin protected service</button>
-    </div>
-
-    <div class="panel panel-default">
-      <div class="panel-heading">
-        <h3 class="panel-title">Response:</h3>
-      </div>
-      <div class="panel-body">
-        <pre id="response"></pre>
-      </div>
-    </div>
-  </div>
-</div>
-
-
-<div class="row">
-  <div id="loggedIn" class="col-md-6">
-    <div class="panel panel-default">
-      <div class="panel-heading">
-        <h3 class="panel-title">Token information</h3>
-      </div>
-      <div class="panel-body" id="loggedInBody"></div>
-    </div>
-  </div>
-</div>
-
+</app-card-page>
