+++ src/client/src/app/page/logout/logout.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import {AuthenticationService} from "../../service/authentication.service";
 
 @Component({
   selector: 'app-logout',
@@ -7,11 +8,11 @@ import { Component, OnInit } from '@angular/core';
 })
 export class LogoutComponent implements OnInit {
 
-  constructor() { }
+  constructor(private authenticationService: AuthenticationService) { }
 
   ngOnInit() {
 
-    console.log("here");
+    this.authenticationService.logout()
   }
 
 }
