+++ src/client/src/app/partial/card-page/card-page.component.html
@@ -1,10 +1,5 @@
 <style>
-  .card {
-    margin-top   : 3%;
-    margin-bottom: 3%;
-    margin-left  : 30%;
-    margin-right : 30%;
-  }
+
 </style>
 
 <div class="card page-panel">
