+++ src/client/src/app/partial/card-page/card-page.component.scss
@@ -1,0 +1,8 @@
+
+.card {
+
+  margin-top   : 3%;
+  margin-bottom: 3%;
+  margin-left  : 30%;
+  margin-right : 30%;
+}
