+++ src/client/src/app/partial/menu/menu.component.html
@@ -1,48 +1,54 @@
-<nav class="navbar navbar-expand-md navbar-dark bg-dark">
+<div *ngIf="authGuard.loggedIn()">
 
-  <div class="navbar-collapse collapse w-100 order-1 order-md-0 dual-collapse2">
+  <nav class="navbar navbar-expand-md navbar-dark bg-dark">
 
-    <ul class="navbar-nav mr-auto">
+    <div class="navbar-collapse collapse w-100 order-1 order-md-0 dual-collapse2">
 
-      <li class="nav-item active">
-        <a class="nav-link" href="/">Home<span class="sr-only">(current)</span></a>
-      </li>
+      <ul class="navbar-nav mr-auto">
 
-      <li class="nav-item">
-        <a class="nav-link" href="/inventory">Inventory</a>
-      </li>
-    </ul>
+        <li class="nav-item active">
+          <a class="nav-link" href="/">Home<span class="sr-only">(current)</span></a>
+        </li>
 
-  </div>
+        <li class="nav-item">
+          <a class="nav-link" href="/inventory">Inventory</a>
+        </li>
+      </ul>
 
-  <div class="mx-auto order-0">
+    </div>
 
-    <a class="navbar-brand mx-auto" href="#">Point of Sale</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target=".dual-collapse2">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-  </div>
+    <div class="mx-auto order-0">
 
-  <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
+      <a class="navbar-brand mx-auto" href="#">Point of Sale</a>
 
-    <ul class="navbar-nav ml-auto">
+      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target=".dual-collapse2">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+    </div>
 
-      <li class="nav-item">
-        <a class="nav-link" href="/users">Manage Users</a>
-      </li>
+    <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
 
-      <li class="nav-item">
-        <a class="nav-link" href="/sales">Sales</a>
-      </li>
+      <ul class="navbar-nav ml-auto">
 
-      <li class="nav-item">
-        <a class="nav-link" href="/settings">Settings</a>
-      </li>
+        <li class="nav-item">
+          <a class="nav-link" href="/users">Manage Users</a>
+        </li>
 
-      <li class="nav-item">
-        <a class="nav-link" href="/logout">Logout</a>
-      </li>
-    </ul>
-  </div>
+        <li class="nav-item">
+          <a class="nav-link" href="/sales">Sales</a>
+        </li>
 
-</nav>
+        <li class="nav-item">
+          <a class="nav-link" href="/settings">Settings</a>
+        </li>
+
+        <li class="nav-item">
+          <a class="nav-link" href="/logout">Logout</a>
+        </li>
+      </ul>
+    </div>
+
+  </nav>
+
+</div>
+
