+++ src/client/src/app/partial/two-page/two-page.component.html
@@ -16,7 +16,6 @@
         'col-md-9'  : left == 9,
         'col-md-10' : left == 10,
         'col-md-11' : left == 11
-
       }">
 
         <div class="col-md-12 page-panel">
@@ -37,7 +36,6 @@
         'col-md-9'  : right == 9,
         'col-md-10' : right == 10,
         'col-md-11' : right == 11
-
       }">
 
         <div class="col-md-12 page-panel">
