--- src/client/src/app/service/session.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SessionService } from './session.service';
-
-describe('SessionService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: SessionService = TestBed.get(SessionService);
-    expect(service).toBeTruthy();
-  });
-});
