--- src/client/src/app/service/session.service.ts
@@ -1,16 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class SessionService {
-
-  constructor() { }
-
-  static logout(){
-
-    console.log("Logged out?");
-  }
-
-
-}
